app/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Alexandre Henrique Afonso Campos <camposalexandreh@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.104.1"
uvicorn = {extras = ["standard"], version = "^0.24.0.post1"}
sqlalchemy = "^2.0.23"
python-creditcard = {git = "https://github.com/maistodos/python-creditcard.git", rev = "main"}
python-dateutil = "^2.8.2"
python-dotenv = "^1.0.0"
psycopg2-binary = "^2.9.9"


[tool.poetry.group.dev.dependencies]
black = "^23.11.0"
pytest = "^7.4.3"
httpx = "^0.25.2"
parameterized = "^0.9.0"
pytest-dotenv = "^0.5.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
